requirements.txt
discord.py>=2.3.2
yfinance>=0.2.36
pandas>=2.0.3
requests>=2.31.0
python-dotenv>=1.0.0